Mon, May 29, 2023, 8:24 AM +00:00

+++ phonebook.c
@@ -6,8 +6,8 @@ int main(void)
 {
     string name = get_string("Name of the person: ");
     int age = get_int("Age of the person: ");
-    double phone = get_int("Phone number: ");
+    long phone = get_int("Phone number: ");
 
 
-    printf("name= %s, age= %i, phone number= %f", name, age, phone);
+    printf("name= %s, age= %i, phone number= %li", name, age, phone);
 }
